Do not send Pause/Resume to API if we know it will be rejected
The "Resume/Stop charging" buttons are disabled/grayed out when they cannot be used.

I tried doing the same for the "Charging"-switch, but it appears it just ignores the availability setting and triggers anyway. The is_command_valid check in the command-function will still block the API call if we are in a state where the command is not legal, and raising the exception triggers a popup in the UI for direct user feedback.