Remove internal circuits api calls (#186)
* Remove internal circuits api calls
* Remove Circuit class in api
* Automatically remove Circuit-devices and their associated entities in HA
* Update changelog+readme

The installation/{self.installation.id}/hierarchy api contains a list of circuits associated with an installation which contains the same information as the circuits/{id} api. Set the Chargers as directly connected to the Installation, since the Circuit device had very limited usability. The circuit information is available in the charger attributes.