Migrate fully to pydantic 2
Fixes #267

There were a couple of attributes that were commented out for some reason. Have set these as optional (`type | None = None`) for the time being, though I'm not sure why these were commented out in the first place.

Should we make these attributes (and/or others) required in addition to the ones we already require, or should we leave it as-is?