Cleanup main readme and reduce duplication of release notes
With 0.8.2 having been out for over a month with little to no (self-created) issues, I think we can replace the the "not extensively tested"-warning-sign with a more generic "we are only able to test so much"-warning.

There was also a lot of duplication of the release notes that I think can be removed (but open to discuss it).